flesh out benchmark

+++ src/test/resources/example-pgbench/V1__example-pgbench.sql
@@ -1,28 +1,28 @@
 CREATE TABLE pgbench_history (
-  tid int,
-  bid int,
-  aid int,
+  tid bigint,
+  bid bigint,
+  aid bigint,
   delta int,
   mtime timestamp,
   filler char(22)
 );
 
 CREATE TABLE pgbench_tellers (
-  tid int not null,
-  bid int,
+  tid bigint not null,
+  bid bigint,
   tbalance int,
   filler char(84)
 ) WITH (fillfactor=100);
 
 CREATE TABLE pgbench_accounts (
-  aid int not null,
-  bid int,
+  aid bigint not null,
+  bid bigint,
   abalance int,
   filler char(84)
 ) WITH (fillfactor=100);
 
 CREATE TABLE pgbench_branches (
-  bid int not null,
+  bid bigint not null,
   bbalance int,
   filler char(88)
 ) WITH (fillfactor=100);
